Add files via upload
Present day data sets. 
MercedHI_Q_T_20220223.txt -- This dataset is discharge and water temperature data from the Merced River at Happy Isles gaging station. This is a long record. The discharge data (Q) goes back to 1915 and the temperature dataset goes back to the 1960s. Data were queried from NWIS on 2/23/2022

MFSanJoaquin_Q_20220223 - This is discharge data from the Middle Fork of the San Joaquin in Devils Postpile gage. This is a newer gage that was installed in 2009. It only has discharge. We recently added temperature, but that is a very new record. Data were queried from NWIS on 2/23/2022.